fix: Fix the logout function and secure login connections

+++ backend/users/apis.py
@@ -1,7 +1,12 @@
-from rest_framework import views, response, permissions #, exceptions
+from rest_framework import views, response, permissions, parsers, renderers, status #, exceptions
 from .serializers import UserSerializer
 from . import services
-
+from rest_framework.request import Request
+from rest_framework.authtoken.serializers import AuthTokenSerializer
+import jwt
+from datetime import datetime
+from django.conf import settings
+from rest_framework.response import Response
 
 class RegisterUserApi(views.APIView):
 
@@ -42,7 +47,7 @@ class LoginUserApi(views.APIView):
         password = request.data.get("password")
         
         try:
-            user = services.login_user(email, password)
+            user = services.login_user(request, email, password)
             # new user accounts should be validated by the admin before being able tologin
             if not user.is_staff:
                 return response.Response(data={"error": "Unauthorized"}, status=401)
@@ -59,12 +64,16 @@ class LoginUserApi(views.APIView):
 class LogoutUserApi(views.APIView):
     authentication_classes = (services.ScriberUserAuthentication,)
     permission_classes = (permissions.IsAuthenticated,)
+    # permission_classes = (services.ScriberUserAuthentication,)
 
-    def post(self, request):
+    def post(self, request: Request):
+        if not request.user.is_authenticated:
+            return response.Response({"message": "User not authenticated"}, status=403)
+
         response_ = response.Response({"message": "Logout successed"},status=200)
         response_.delete_cookie("jwt")
         user_email = request.user.email
-        services.logout_user(user_email)
+        services.logout_user(request, user_email)
         return response_
 
 class UserApi(views.APIView):
